mbta2 db; emergency rss feed
there are changes to the alert RSS feed as part of the move to
MBTA-realtime. Unfortunately we are no longer able to have a
backwards-compatible feed at the old URL as originally planned. *Any
developer using an RSS feed of alerts must switch to the new URL right
away. *
*
*
The backward-compatible feed is available at:
http://realtime.mbta.com/alertsrss/rssfeed2
This replicates the format of the old feed. We do not expect to support it
indefinitely.

The new-and-improved format feed is available at:
http://realtime.mbta.com/alertsrss/rssfeed4

More information about the two feeds can be found in the MBTA-realtime
documentation.  <http://realtime.mbta.com/Portal/Home/Documents>

We apologize for the lack of notice.

Sincerely,
developer at MBTA